+++ 3495/2/commands/checkin_command/checkin.rb
@@ -1,5 +1,6 @@
 module CheckinCommand
   TIME_STAMP = Time.now.strftime('%d/%m/%Y %H:%M').tr('/', '.')
+  API_URL = 'https://api.telegram.org/'.freeze
   def checkin!(*)
     check_sign_up
   end
@@ -50,7 +51,7 @@ module CheckinCommand
   end
 
   def request_file_path(file_id)
-    uri = URI("#{API_URL}bot886244897:AAE8balNKJ7Nukdam2v3AuhiAhxCyRysVBs/getFile?file_id=#{file_id}")
+    uri = URI("#{API_URL}bot#{ENV['TOKEN']}/getFile?file_id=#{file_id}")
     json_response = JSON.parse(Net::HTTP.get(uri))
     load_pic_from_path(json_response['result']['file_path'])
   end
@@ -62,10 +63,10 @@ module CheckinCommand
   end
 
   def load_pic_from_path(file_path)
-    uri = URI("#{API_URL}file/bot886244897:AAE8balNKJ7Nukdam2v3AuhiAhxCyRysVBs/#{file_path}")
+    uri = URI("#{API_URL}file/bot#{ENV['TOKEN']}/#{file_path}")
     DirCreator.dir_create("public/#{from['id']}/checkin/#{TIME_STAMP}")
     photo_new_path = "public/#{from['id']}/checkin/#{TIME_STAMP}/selfie.jpg"
-    File.write(photo_new_path, Kernel.open(uri), mode: 'wb')
+    File.write(photo_new_path, Kernel.open(uri).read, mode: 'wb')
     photo_new_path
   end
 
@@ -73,5 +74,4 @@ module CheckinCommand
     (53.914264..53.916233).cover?(location['latitude'].to_f) &&
       (27.565941..27.571306).cover?(location['longitude'].to_f)
   end
-  API_URL = 'https://api.telegram.org/'.freeze
 end
