+++ 3495/2/lib/webhook_controller.rb
@@ -1,0 +1,7 @@
+class WebhooksController < Telegram::Bot::UpdatesController
+  include Telegram::Bot::UpdatesController::Session
+  include Telegram::Bot::UpdatesController::MessageContext
+  include StartCommand
+  include CheckinCommand
+  include CheckoutCommand
+end
