--- 3495/2/public/328230825/checkin/22.07.2019 10:34/selfie.jpg
@@ -1,1 +1,0 @@
-#<File:0x000055f3e84ad030>
