--- 3495/2/public/328230825/checkin/22.07.2019 11:17/selfie.jpg
@@ -1,1 +1,0 @@
-#<File:0x000055839a1dd5f8>
