--- 3495/2/public/328230825/checkout/22.07.2019 11:17/geo.txt
@@ -1,1 +1,0 @@
-53.914374, 27.569301
