--- 3495/2/public/328230825/checkout/22.07.2019 11:17/selfie.jpg
@@ -1,1 +1,0 @@
-#<File:0x000055839a302960>
