+++ 3495/2/telegram_bot.rb
@@ -12,18 +12,12 @@ require_relative './commands/start_command/start.rb'
 require_relative './commands/checkin_command/checkin.rb'
 require_relative './commands/checkout_command/checkout.rb'
 require_relative './lib/dir_creator.rb'
+require_relative './lib/webhook_controller.rb'
 Dotenv.load
 Ohm.redis = Redic.new('redis://127.0.0.1:6379')
 Telegram::Bot::UpdatesController.session_store = :redis_store, { expires_in: 1.year }
-class WebhooksController < Telegram::Bot::UpdatesController
-  include Telegram::Bot::UpdatesController::Session
-  include Telegram::Bot::UpdatesController::MessageContext
-  include StartCommand
-  include CheckinCommand
-  include CheckoutCommand
-end
 
-TOKEN = '886244897:AAE8balNKJ7Nukdam2v3AuhiAhxCyRysVBs'.freeze
+TOKEN = ENV['TOKEN'].freeze
 bot = Telegram::Bot::Client.new(TOKEN)
 logger = Logger.new(STDOUT)
 poller = Telegram::Bot::UpdatesPoller.new(bot, WebhooksController, logger: logger)
